+++ test/test_helper.rb
@@ -1,9 +1,23 @@
-$LOAD_PATH.unshift File.expand_path("../../lib", __FILE__)
-require "minidust"
+require 'minitest/autorun'
+require 'minitest/pride'
+require 'stringio'
+
+$LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
+require 'minidust'
 
 # Minidust.enable!
 
 # require 'minitest/mock'
 
-require "minitest/autorun"
 require 'mocha/minitest'
+
+class MinitestWithStdoutCapture < Minitest::Test
+  def capture_stdout
+    old_stdout = $stdout
+    $stdout = StringIO.new
+    yield
+    $stdout.string
+  ensure
+    $stdout = old_stdout
+  end
+end