+++ Changelog.md
@@ -37,3 +37,9 @@ All notable changes to this project will be documented in this file.
 - Update README
 - Added a default minidust.yml - you can override which paths are to be included and excluded in your coverage
 - Added unused methods to minidust coverage report
+- Add emojis to coverage report output
+
+
+## [0.1.7] - 2025-06-04
+### Added
+- Update README